requirements.txt
langchain
openai
pinecone-client
langchain-pinecone
langchain-openai
python-dotenv
pypdf
langchain_community
langchain-chroma
langchain-ollama
jq
text_generation
gradio